extract derivedLinksTypes, fix subscription race and stale nested entries bugs, deduplicate startLinkObservation
